JAMES-2341 Move testing tool on MimeMessage to a common project
